refactor(comfybuilder): deduplicate JWT-claims decoding, simplify pkce and loopback
- one statusFromAccessToken in claims.ts replaces the two drifting
  decoders in oauth.ts and tokenStore.ts; a malformed token after a
  successful exchange now degrades to unset identity fields instead of
  throwing
- pkce uses Node's native base64url encoding
- loopback's three identical failure responses collapse into failRequest
